Add import_metadata (python) and import_metadata.sh (bash) scripts
These two scripts both serve the same purpose in that they can take the data
produced by *classify* to update the database and/or produce SQL to do so at a
later moment. The python version is much faster at parsing CSV data and
updating the database, for the rest the scripts are functionally identical.